Rearrange gateway `bridge` module into `sharding` and spruce up docs
