--- src/gateway/bridge/event.rs
@@ -1,20 +1,0 @@
-//! A collection of events created by the client, not a part of the Discord API itself.
-
-use crate::gateway::ConnectionStage;
-use crate::model::id::ShardId;
-
-/// An event denoting that a shard's connection stage was changed.
-///
-/// # Examples
-///
-/// This might happen when a shard changes from [`ConnectionStage::Identifying`] to
-/// [`ConnectionStage::Connected`].
-#[derive(Clone, Debug)]
-pub struct ShardStageUpdateEvent {
-    /// The new connection stage.
-    pub new: ConnectionStage,
-    /// The old connection stage.
-    pub old: ConnectionStage,
-    /// The ID of the shard that had its connection stage change.
-    pub shard_id: ShardId,
-}
