+++ src/gateway/client/mod.rs
@@ -2,7 +2,24 @@
 //! events to handlers and starting sharded gateway connections is handled directly by the client.
 //! In addition, the client automatically handles caching via the [`Cache`] struct.
 //!
+//! # Sharding
+//!
+//! If you do not require sharding - such as for a small bot - then use [`Client::start`]. If you
+//! don't know what sharding is, refer to the [`sharding`] module documentation.
+//!
+//! There are a few methods of sharding available:
+//! - [`Client::start_autosharded`]: retrieves the number of shards Discord recommends using from
+//!   the API, and then automatically starts that number of shards.
+//! - [`Client::start_shard`]: starts a single shard for use in the instance, handled by the
+//!   instance of the Client. Use this if you only want 1 shard handled by this instance.
+//! - [`Client::start_shards`]: starts all shards in this instance. This is best for when you want a
+//!   completely shared State.
+//! - [`Client::start_shard_range`]: start a range of shards within this instance. This should be
+//!   used when you, for example, want to split 10 shards across 3 instances.
+//!
 //! Click [here][Client#examples] for an example on how to use a [`Client`].
+//!
+//! [`sharding`]: crate::gateway::sharding
 
 mod context;
 pub(crate) mod dispatch;
@@ -357,7 +374,7 @@ impl IntoFuture for ClientBuilder {
 
 /// The Client is the way to be able to start sending authenticated requests over the REST API, as
 /// well as initializing a WebSocket connection through [`Shard`]s. Refer to the [documentation on
-/// using sharding][sharding docs] for more information.
+/// using sharding][super::sharding] for more information.
 ///
 /// # Event Handlers
 ///
@@ -399,7 +416,6 @@ impl IntoFuture for ClientBuilder {
 ///
 /// [`Shard`]: crate::gateway::Shard
 /// [`Event::MessageCreate`]: crate::model::event::Event::MessageCreate
-/// [sharding docs]: crate::gateway#sharding
 pub struct Client {
     data: Arc<dyn std::any::Any + Send + Sync>,
     /// A HashMap of all shards instantiated by the Client.
