+++ src/gateway/sharding/shard_queuer.rs
@@ -10,13 +10,10 @@ use tokio::sync::Mutex;
 use tokio::time::{sleep, timeout, Duration, Instant};
 use tracing::{debug, info, warn};
 
-#[cfg(feature = "voice")]
-use super::VoiceGatewayManager;
 use super::{
     ShardId,
     ShardManager,
     ShardMessenger,
-    ShardQueuerMessage,
     ShardRunner,
     ShardRunnerInfo,
     ShardRunnerOptions,
@@ -25,6 +22,8 @@ use super::{
 use crate::cache::Cache;
 #[cfg(feature = "framework")]
 use crate::framework::Framework;
+#[cfg(feature = "voice")]
+use crate::gateway::VoiceGatewayManager;
 use crate::gateway::{
     ConnectionStage,
     InternalEventHandler,
@@ -341,4 +340,17 @@ impl ShardQueue {
     pub fn buckets_filled(&self) -> bool {
         self.buckets.iter().all(|b| !b.is_empty())
     }
+}
+
+/// A message to be sent to the [`ShardQueuer`].
+#[derive(Clone, Debug)]
+pub enum ShardQueuerMessage {
+    /// Message to set the shard total.
+    SetShardTotal(NonZeroU16),
+    /// Message to start a shard.
+    Start { shard_id: ShardId, concurrent: bool },
+    /// Message to shutdown the shard queuer.
+    Shutdown,
+    /// Message to dequeue/shutdown a shard.
+    ShutdownShard { shard_id: ShardId, code: u16 },
 }
